Create swap-terminal for stables coins with time and price trigger
Consider a costumer wants to swap assets t stable coins, no difference for USDC/USDT/ or any other native USD coin, just a stable coin. 
what if the rates are not what he wants. creates an order, sets time "X" minutes, and the desired rate! that's it. 
If the conditions meet each of stable coins, the transaction would be done. 
The Done is Done. :)